+++ libOctree.c
@@ -2,7 +2,7 @@
 #include "stdlib.h"
 #include "string.h"
 #include "math.h"
-#include "stdint.h"
+#include "inttypes.h"
 #include "tools.h"
 #include "libOctree.h"
 
@@ -139,7 +139,7 @@ void mapOctree(int level,octreeStruct *octree,treeStruct **tree,double x,double 
     }else{  /*this is the lowest level, allocate map space*/
       if(octree->mapFile){
         if(!(octree->mapFile=(int **)realloc(octree->mapFile,(octree->nMaps+1)*sizeof(int *)))){
-          fprintf(stderr,"Error in octree reallocation, %lu\n",(octree->nMaps+1)*sizeof(int *));
+          fprintf(stderr,"Error in octree reallocation, %" PRIu64 "\n",(uint64_t)(octree->nMaps+1)*sizeof(int *));
           exit(1);
         }
       }else octree->mapFile=iIalloc(octree->nMaps+1,"mapFile",0);
